fix: :file_folder: Renamed component filename to UpperCamelCase
Issue: #152
