+++ src/components/GameList/index.tsx
