+++ src/components/NewsList/index.tsx
