+++ src/components/PachiBorder/index.tsx
